Add builtin function names
When you print a function object, it will now also print its name!

Implemented by storing a name field in lval...not sure if that's the
best way but otherwise a LVAL_FUN doesn't know its own name. Not sure if
I should have overloaded the sym field and just used that since it's not
used in functions...